backend/pyproject.toml
[tool.poetry]
name = "odb-lab-backend"
version = "0.1.0"
description = ""
authors = ["fooqoo <fooqoofooqoo56@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
open-interpreter = "^0.3.12"
fastapi = "^0.114.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
black = "^23.12.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
